requirements.txt
fake-factory
logbook
pillow
